requirements.txt
streamlit==1.32.2
google-generativeai
pillow==10.2.0
pypdf==4.0.1
python-docx==1.1.0
fpdf2==2.7.8
